BOAC-1808, inactive-asc and inactive-coe constraints removed from /api/search
