+++ src/api/student.ts
@@ -19,19 +19,15 @@ export function getStudent(uid: string) {
 export function search(
   phrase: string,
   includeCourses: boolean,
-  isInactiveAsc: boolean,
-  isInactiveCoe: boolean,
   orderBy: string,
   offset: number,
   limit: number
 ) {
   let apiBaseUrl = store.getters['context/apiBaseUrl'];
   return axios
-    .post(`${apiBaseUrl}/api/students/search`, {
+    .post(`${apiBaseUrl}/api/search`, {
       searchPhrase: phrase,
       includeCourses: includeCourses,
-      isInactiveAsc: isInactiveAsc,
-      isInactiveCoe: isInactiveCoe,
       orderBy: orderBy || 'first_name',
       offset: offset || 0,
       limit: limit || 50
